+++ src/sn_network.h
@@ -356,7 +356,7 @@ int sn_network_serialize (sn_network_t *n, char **ret_buffer,
  */
 sn_network_t *sn_network_unserialize (char *buffer, size_t buffer_size);
 
-uint32_t sn_network_get_hashval (const sn_network_t *n);
+uint64_t sn_network_get_hashval (const sn_network_t *n);
 
 #endif /* SN_NETWORK_H */
